step_3(incompelete)/requirements.txt
kafka-python
requests
